+++ examples/mdns-passive-discovery.rs
@@ -31,7 +31,7 @@ async fn main() -> Result<(), Box<dyn Error>> {
     println!("Local peer id: {:?}", peer_id);
 
     // Create a transport.
-    let transport = libp2p::build_development_transport(id_keys)?;
+    let transport = libp2p::development_transport(id_keys).await?;
 
     // Create an MDNS network behaviour.
     let behaviour = Mdns::new(MdnsConfig::default()).await?;
@@ -40,7 +40,7 @@ async fn main() -> Result<(), Box<dyn Error>> {
     // Note that the MDNS behaviour itself will not actually inititiate any connections,
     // as it only uses UDP.
     let mut swarm = Swarm::new(transport, behaviour, peer_id);
-    Swarm::listen_on(&mut swarm, "/ip4/0.0.0.0/tcp/0".parse()?)?;
+    swarm.listen_on("/ip4/0.0.0.0/tcp/0".parse()?)?;
 
     loop {
         match swarm.next().await {