+++ CHANGELOG.md
@@ -1,4 +1,3 @@
-
 # Individual crates
 
 ## Main APIs
@@ -42,7 +41,26 @@
 
 # `libp2p` facade crate
 
-## Version 0.36.0 [unreleased]
+## Version 0.37.0 [unreleased]
+
+- Update individual crates.
+    - `libp2p-floodsub`
+    - `libp2p-gossipsub`
+    - `libp2p-kad`
+    - `libp2p-mdns`
+    - `libp2p-ping`
+    - `libp2p-relay`
+    - `libp2p-request-response`
+    - `libp2p-swarm`
+
+## Version 0.36.0 [2021-03-17]
+
+- Consolidate top-level utility functions for constructing development
+  transports. There is now just `development_transport()` (available with default features)
+  and `tokio_development_transport()` (available when the corresponding tokio features are enabled).
+  Furthermore, these are now `async fn`s. The minor variations that also included `pnet`
+  support have been removed.
+  [PR 1927](https://github.com/libp2p/rust-libp2p/pull/1927)
 
 - Update libp2p crates.