+++ protocols/floodsub/CHANGELOG.md
@@ -1,4 +1,8 @@
-# 0.28.0 [unreleased]
+# 0.29.0 [unreleased]
+
+- Update `libp2p-swarm`.
+
+# 0.28.0 [2021-03-17]
 
 - Update `libp2p-swarm`.